requirements.txt
django
djangorestframework
django-cors-headers
requests
psycopg2-binary
python-decouple
pytest
pytest-django
pytest-cov
